refactor: move db startup
