+++ src/TaskList.Presentation/Program.cs
@@ -1,7 +1,7 @@
-using Microsoft.EntityFrameworkCore;
 using TaskList.Domain.Repositories;
 using TaskList.Persistence;
 using TaskList.Persistence.Repositories;
+using TaskList.Presentation.StartUp;
 using TaskList.Presentation.Middlewares;
 using TaskList.Services;
 using TaskList.Services.Abstractions;
@@ -12,7 +12,7 @@ builder.Services.AddControllers();
 builder.Services.AddSwaggerGen();
 builder.Services.AddScoped<IServiceManager, ServiceManager>();
 builder.Services.AddScoped<IRepositoryManager, RepositoryManager>();
-builder.Services.AddDbContext<RepositoryDbContext>(SetConnectionString);
+builder.Services.AddDbContext<RepositoryDbContext>(DataBaseStartUp.SetConnectionString);
 builder.Services.AddTransient<ExceptionHandlingMiddleware>();
 
 var app = builder.Build();
@@ -21,24 +21,6 @@ app.UseSwagger();
 app.UseSwaggerUI();
 app.MapControllers();
 
-await ApplyMigrations(app.Services);
+await DataBaseStartUp.ApplyMigrations(app.Services);
 
 app.Run();
-
-static async Task ApplyMigrations(IServiceProvider serviceProvider)
-{
-    using var scope = serviceProvider.CreateScope();
-    await using var dbContext = scope.ServiceProvider.GetRequiredService<RepositoryDbContext>();
-    await dbContext.Database.EnsureCreatedAsync();
-    await dbContext.Database.MigrateAsync();
-}
-
-static void SetConnectionString(DbContextOptionsBuilder dbContextOptionsBuilder)
-{
-#if RELEASE
-    var connectionString = Environment.GetEnvironmentVariable("dbConnectionString");
-#else
-    const string connectionString = "Host=127.0.0.1;Port=5432;Database=task_list_db;Username=postgres_user;Password=postgres_pwd";
-#endif
-    dbContextOptionsBuilder.UseNpgsql(connectionString);
-}
